Initial project file
We use Maven (https://maven.apache.org/) as build management tool in this project.

Maven will allow us to compile and run our server application in a reproducible and consistent way.

We will use the latest Java version 17 with long-term support (LTS).

These decisions (and others) are configured in the pom.xml file.

> POM stands for `Project Object Model` and contains information about the project and configuration details used by Maven to build the project.
(Source: Introduction to the POM, https://maven.apache.org/guides/introduction/introduction-to-the-pom.html)

**tags**: #NotRelevantForExam, codingSession01start, GeneralProjectSetup