Add supporting Idea project files and gitignore
We commit several Idea project files that should be the same for everybody.

We also exclude some Idea files form version control, because they are specific to the computer the application is build, e.g. they contain the username in some file paths, that are different for everybody.
We ignore the build directory of Maven (`target`) in git.

We will update these files from time to time during the lecture.

**tags**: #NotRelevantForExam