Add build dependency for creating an executable jar
We add a build dependency so that we can create an executable jars (sometimes called “fat jars”) are archives containing your compiled classes along with all the jar dependencies that your code needs to run.

The build dependency also adds custom build goals, that starts the server in a blocking or unblocking way.

Additional Resources:
 - https://docs.spring.io/spring-boot/docs/current/reference/html/getting-started.html#getting-started.first-application.executable-jar
 - https://docs.spring.io/spring-boot/docs/2.7.6/maven-plugin/reference/htmlsingle/#goals

**tags**: #NotRelevantForExam