Add simple date endpoint to /
We add the first functionality to our application.
We will return the current date and time.

For this we have to map the URL http://localhost:8080/ to a Java method.
We use the annotation `RequestMapping` for this configuration.
Additionally, we have to add the annotation `RestController` to our `Application` class.

Both annotations will be present during runtime of our application, and configure the Spring Boot application server.

## Experiments
- Change the return value to a different String.
- Change the return type to a different Java type (long, Date, ...)
- Change the String "/" of the `RequestMapping`
- See the results in your browser (Only change one thing at a time!).

**tags**: #RelevantForExam, AddFirstSimpleServerLogic