Added sentence-split method and endpoint.
Added the sentence-split method for exercise 1.2.

We create the Java method `splitSentence` with the String parameter `sentenceJavaMethodParameter`.
This method will be offered as new endpoint via the URL (http://localhost:8080/sentence-split). For this we add the `RequestMapping` to our new method with the path `/sentence-split`.
We also map the request parameter `url_parameter_name_sentence` to our Java method parameter `sentenceJavaMethodParameter`.
This will use any value assigned to the name `url_parameter_name_sentence` in the HTTP request and pass it as `sentenceJavaMethodParameter` into our Java method `splitSentence`.

As a return value, we build a simple html page that list the split words as ordered list `<ol>` in html.

**tags**: #RelevantForExam