Added sentence HTML form
Added the sentence-split form for exercise 1.2.

See the different names for the sentence variables to get an understanding which names belong together:
- html_id_sentence
- url_parameter_name_sentence
- sentenceJavaMethodParameter

Whenever you change one of the names in one place, you must also change the other place.
If you forget one place you will get the errors at different times during development/execution:
- html_id_sentence:
  - Browser Developer Tools
  - IDEA Problems view
- url_parameter_name_sentence:
  - Runtime Error after submit in Browser
  - Logged warning (WARN) in server output that "Required request parameter is [...] not present"
- sentenceJavaMethodParameter:
  - IDEA error "Cannot resolve symbol `sentenceJavaMethodParameter`"
  - Compiler error "cannot find symbol"

This is another example of the difficulty of client-server development.
The programming of the client server mapping "url_parameter_name_sentence" is only visible at runtime, which makes this type of error difficult to find during development.
You have to test the integration of client and server in order to find this kind of error.

**tags**: #RelevantForExam