Add service class for reviews
We add a `ReviewService` class that should hold all our review business logic.

We use the `Service` annotation to signal the business layer in this class to the framework and other developers.
Currently, there is no difference between the generic Spring annotation `Component` and `Service`.
But in later future releases it could add additional features only relevant for the business layer.

Additional Resources:
- https://docs.spring.io/spring-framework/docs/current/reference/html/core.html#beans-stereotype-annotations
- https://stackoverflow.com/a/6897038/2042627

**tags**: #RelevantForExam