Add the ReviewService as dependency to the ReviewController
We add the new ReviewService as dependency to the ReviewController.

To configure how the Spring Boot framework inject a reviewService instance to the ReviewController, we use the annotation @Autowired at the constructor of the ReviewController.

Quote from the documentation (https://docs.spring.io/spring-framework/docs/5.3.x/reference/html/core.html#beans-constructor-injection):

> ... it is a good rule of thumb to use constructors for mandatory dependencies and setter methods or configuration methods for optional dependencies.

**tags**: #RelevantForExam