Change endpoint get single review to path variable
We change the place of the index parameter in our URL.
We move it from the query part into the path part of the URL.

This endpoint pattern is good practice, following the Representational State Transfer (REST) paradigm.

For us in the lecture, it is important, that we can use different parts of the HTTP request to map input data to Java method parameters.

**tags**: #RelevantForExam