Add Review class
In our review domain a review consists of a comment, a rating and a submission date.

In our server application we want to change from the simple review comment list to the full object structure.
Therefore, we add the review POJO (Plain Old Java Object) with the necessary fields and methods.
We also add a constructor for this class.

Notice that we only use the comment and rating as parameters in the constructor.

**tags**: #RelevantForExam