Add ReviewDetails
Add ReviewDetails Page as additional Page.

Change Link in ReviewList.html to relative link of ReviewDetails.html with index as fragment URL part.

We use the `onload` event of the `body` to execute the `loadReviews()` function, so that we don't need to click the button in order to load the review.

**tags**: #RelevantForExam