Configure the UserController as Spring Boot  @RestController
The annotation @RestController allows the Spring Boot Framework to recognize the UserController Java class as HTTP request handler, that has request mapping rules configured for the Java methods of this class.

Because there are no methods in this class at the moment nor are there defined request mappings, nothing will change, if you start the application.

**tags**: #RelevantForExam