Add getUsers Java method
We add a simple Java method that will return all users of the application.

At the moment we will just return an empty List of Strings.
It is just enough code, so that the Class can be complied without errors.

If you start the application, it will have no new documentation under (http://localhost:8080/swagger-ui.html); as there is still missing the @RequestMapping annotation.

**tags**: #RelevantForExam