Add GET RequestMapping to getUsers method
We add the @GetMapping annotation to the method.

This will register the path "/users" for the HTTP method "GET" to the getUser method in the application.

Now HTTP GET requests (http://localhost:8080/users) will return a successful HTTP response.
At the moment only an empty list `[]` is returned, because our Java method returns only an empty ArrayList.

Before this change, an HTTP GET request to (http://localhost:8080/users) returned a client error (type= Not Found, status= 404).

**tags**: #RelevantForExam