Add getUser Java method
We add the implementation for the feature "Show user with specific Id".

Again the implementation is not yet mapped to specific HTTP requests in our application.
So it cannot be used from clients, nor is it shown in our automated documentation under (http://localhost:8080/swagger-ui.html).

Because we did a refactoring in the last commit, we can now easily use the users list also in our new method.

**tags**: #RelevantForExam