Add RequestMapping to getUser Method
We add the @GetMapping annotation to the method.

This will register the path "/users/{id}" for the HTTP method GET to the getUser method in the application.
With the `{id}` we also declare a path variable, that can also be used to send data from the client to the server.

!Important!
Now HTTP GET requests (http://localhost:8080/users/0) will return a server error HTTP response.
The reason for this internal server error can be seen in the console log of the application.

The HTTP request is mapped by the Spring Boot framework to the right Java method, but as this method requires a method parameter "int id", the framework cannot fill this parameter with a value and raises an Exception.

Intellij also raises a warning at the `getUser` method that the declared path variable `id` is not consumed.

**tags**: #RelevantForExam