Add submitUserObject Java method
We add the implementation for the feature "Create new user with name and email address".

At the moment we will just use the given name, and add the name to our users list.
We will ignore the email for now, but we want to have the parameter from the beginning in our method signature.
In this way clients that will use this method have to implement the email parameter right from the start.
If we had added the email parameter later, existing clients would get errors from our application.

Our Java implementation will not return anything as the successful call to this method (without any errors/exceptions) is confirmation enough.

**tags**: #RelevantForExam