Add RequestMapping to submitUserObject Method
We add the @PostMapping annotation to the method.

This will register the path "/users/submit" for the HTTP method "POST" to the submitUserObject method in the application.

!Important!

The HTTP POST requests to (http://localhost:8080/users/submit) will get a successful HTTP response.
Also, the Java method parameters are shown in the automated documentation (http://localhost:8080/swagger-ui/index.html#/user-controller/registerUser) though we did not annotate the method parameters with the @RequestParameter annotation.
This is due to the default mapping logic in Spring Boot, see last paragraph in the documentation (https://docs.spring.io/spring-framework/docs/current/reference/html/web.html#mvc-ann-requestparam)

You will also notice that the request that is sent to the application by the "Try it out" button from our automated documentation is attaching the request parameter as query parameters in the URL and not as request parameters in the body.

As this request parameter mapping is dependent on the default mapping logic of the Spring Framework and can change in future versions,
it is recommended to explicitly state the wanted request mapping logic inside your code.

Also, the parameter mapping logic will accept query and form data to convert data into Java parameter values (see: https://docs.spring.io/spring-framework/docs/current/javadoc-api/org/springframework/web/bind/annotation/RequestParam.html).

Additional References:
* https://docs.spring.io/spring-framework/docs/current/reference/html/web.html#mvc-ann-requestparam
* https://docs.spring.io/spring-framework/docs/current/javadoc-api/org/springframework/web/bind/annotation/RequestParam.html

**tags**: #RelevantForExam