Add Service annotation to UserService class
With the @Service annotation we configure that the lifetime of this class with its instances will be managed by the Spring Boot framework.

That means all instances will be created by the framework and other objects that will depend on this class will get an instance by the framework.

It is not necessary to create an instance of the UserService by yourself.

**tags**: #RelevantForExam