Change internal UserService implementation to use  User class instead of String.
We change the internal UserService implementation to use an ArrayList of User instead an ArrayList of String to store our User data.

This is only an intermediate step to replace the former String User data with the User class.

Advanced stream usage to transform the ArrayList users into the old data structure of ArrayList of Strings for the getUsers method. #NotRelevantForExam

**tags**: #RelevantForExam