Add meaningful labels to the registration form
Add labels to our input boxes, so that users know what to fill out.

A label is not only added to the form, we also define for which input field the label is for.
Do not rely on position, but explicit reference the input field.

Read the documentation "How to structure a web form" (https://developer.mozilla.org/en-US/docs/Learn/Forms/How_to_structure_a_web_form) and find the benefits of using the for attribute.

**tags**: #RelevantForExam