Add business logic to prevent duplicate user  registration
We loop through our existing users and compare each email to the email of the new user.

If there is a user with the same email return the method call without adding the user to our ArrayList.

At the moment we do not return any error status, so a client will not be aware of a duplicate email.

**tags**: #RelevantForExam