Add stable and independent id to Review
We add an identifier (id) to our Review class that will not change after the Review is created.

The identifier is an atomic counter that will increment itself, if a new review is created.

The implementation of this counter is thread-safe, meaning that even with a parallel HTTP request submitting new reviews, the counter will not return the same number twice.

!Important!

The thread-safe guarantee is only given in the same JVM.

Additional Resources:
* https://docs.oracle.com/javase/8/docs/api/java/util/concurrent/atomic/AtomicLong.html

**tags**: #RelevantForExam, codingSession06start