Add JPA and H2 dependencies in pom.xml
We add the external dependencies so that we can use persistent relational data in our application.

In this lecture we will use the in-memory database H2, but connections to other databases can be configured similarly.

The advantage of H2 is that it can be installed via maven as external dependency. Other databases require a separate installation and configuration that would be out of scope for this lecture.

Also update the idea configuration for these dependencies.

Additional Resources:
* https://www.baeldung.com/spring-boot-h2-database
* http://www.h2database.com/html/main.html

**tags**: #NotRelevantForExam