Enable web console for H2 database
We enable the H2 database, so that it is started together with the application.

The web console for the database will be available at (http://localhost:8080/h2).

We change the datasource URL configuration to use a fixed JDBC connection string.

Have a look at the server log during the application start.
The "H2ConsoleAutoConfiguration" will tell you the JDBC URL that you must use to connect to your H2 database. It is "jdbc:h2:mem:cbs-database".

This will allow us to log into the database using the H2 console under (http://localhost:8080/h2) with the same JDBC URL after a restart, without copying the unique URL (that is the default behaviour) after each restart.

The default username is "sa" and the password is empty.

**tags**: #NotRelevantForExam