Add JPA annotations for Review
We add the @Entity annotation to our Review class.
This annotation configures the application that a database can be used as persistence for objects of Review.

The JPA Specification demands also the presents of an identifier.
For now, we will use the existing id for that, and add the @Id annotation to our field id.

We add a protected default constructor, so that the JPA implementation can create new objects without being forced to use our custom constructor.

Start your application and have a look at your database.

Try and run the following SQL queries:
* `SELECT * FROM REVIEW`;
* `SHOW COLUMNS FROM REVIEW`;

**tags**: #RelevantForExam