Add ReviewRepository
We add the ReviewRepository to get access to the configured database persistence via Java methods.
We configure the Review class as type for this repository with Long as the type of the id.

As we want the full magic of String Boot we use the JpaRepository Interface and not only the CrudRepository Interface.

All persistent datasources in SpringBoot are always called *Repository.
This name has its origin in the Domain-Driven Design method.
Another name for this kind of classes is Data Access Objects (DAO).

Additional Resources:
- https://spring.io/guides/gs/accessing-data-jpa/
- https://docs.spring.io/spring-data/jpa/docs/current/reference/html/#repositories
- https://de.wikipedia.org/wiki/Domain-driven_Design

**tags**: #RelevantForExam