Add the ReviewRepository as dependency to the  ReviewService
We add the new ReviewRepository as dependency to the ReviewService.
As before, we choose the constructor dependency.

Now we also have to adapt our former JavaHTTPServer implementation, for now we just add a null value to satisfy the compiler.
But later this will lead to a broken JavaHTTPServer during runtime.

**tags**: #RelevantForExam