Use ReviewRepository instead of in-memory ArrayList  data structure
The ReviewRepository offers equal methods to create, read and delete Reviews with the database:
* reviewRepository.save()
* reviewRepository.findAll()
* reviewRepository.findById(id).get()
  * or reviewRepository.findById(id).orElse(null)
* reviewRepository.deleteById(id);

Only the update method has a special implementation.
We first get the current state of the data before the update, then we override only comment and rating, and after that we save the changed state of the data.

First this has business logic reasons, we want to make sure that the submittedAt date stays the same and cannot be changed.

Second it has something to do with the persistence context of an entity.
If the Review object is sent by the client, the object is not in the persistence context.
Therefore, we have to get the real Review entity from the persistence context in order to change some data.

!Import!

The counter of our ids starts now at 1.

Additional Resources:
* https://www.baeldung.com/spring-data-crud-repository-save#updateInstance
* https://www.baeldung.com/hibernate-save-persist-update-merge-saveorupdate#2-states-of-entity-instances

**tags**: #RelevantForExam