Add findAllReviewsSorted method to retrieve the  newest reviews first
In a Repository we add method signatures and not method implementations as the Repository is an interface and not a class.

We use @Query annotations to declare what query in JPQL should be used, when the reviewRepository.findAllReviewsSorted() method is called.

If you use certain keywords in your Java method name, the Spring Boot Framework will translate the method name into a query.
See examples here: https://docs.spring.io/spring-data/jpa/docs/current/reference/html/#jpa.query-methods.query-creation
But this will not work for "findAll" queries.

Sources:
- https://www.baeldung.com/spring-data-jpa-query
- https://docs.spring.io/spring-data/jpa/docs/current/reference/html/#repositories.query-methods.details

**tags**: #RelevantForExam