Add JPA annotation for User
* Add @Entity annotation for the class.
* Add @Table annotation for the class, because of the reserved keyword USER in SQL.
* Add field id of type Long
* Add @Id annotation to field id
* Add @GeneratedValue annotation to field, so that the database will take care of maintaining the id counter.
* Add protected default constructor for ORM
* Add getId() method for mapping to JSON object

Because there are reserved keywords in SQL we cannot create a table with the name "user". The name "user" is reserved for the current user of the database.
We have to use a different name for the table. We can use the table name "USERS" instead.
Use the annotation `@Table(name = "USERS")` at the class `User` to change the table name.

**tags**: #RelevantForExam