Move the initial data setup into separate Java class
Move the data setup for all of our classes into a separate Java class.

Therefore, we define dependencies to all repositories as we already know it in our services.

This will be easier to maintain with the upcoming relations.
Also, our application is not filled with data setup logic.

We also use a logger to signal to us the start and the end of the data initialization, and potential errors between.
See the result in the server logs when you start the server, look for lines with class name `study.coco.csb.server.DataLoader`.

Another option would have been a separate SQL script, more information about this in the additional resources.

Additional Resources:
- https://docs.spring.io/spring-boot/docs/current/reference/html/howto.html#howto-initialize-a-database-using-spring-jdbc
- https://www.baeldung.com/spring-boot-data-sql-and-schema-sql

**tags**: #NotRelevantForExam