Add ManyToOne relation Review to Shop
We add a relation between Review and Shop.
As a Review will always belong to one Shop, and a Shop can have many Reviews, we can choose the ManyToOne or OneToMany relationship depending on the viewpoint we want to take.

Because we add the beginning of the relation into the Review Class, we use the @ManyToOne annotation.
As we will also only add the relationship to the Review it is a unidirectional relationship.

The usage of a unidirectional OneToMany relationship should be avoided.

Remember:
* a @ManyToOne annotation belongs at a simple field, e.g. `Shop`
* a @OneToMany annotation belongs at a list, e.g. `List<Reviews>`

We also adapt our DataLoader class, so that the reviews are already connected to shops.

Experiment:
- Try to delete a shop that has reviews. See the error message in the server logs.
- Delete all the connected reviews.
- Try to delete the shop again. Does it work now?

Sources:
- https://thorben-janssen.com/best-practices-many-one-one-many-associations-mappings/

**tags**: #RelevantForExam