Add OneToOne relation User to Shop
We add a relation between User and Shop.
As a User will belong to one Shop, and a Shop can have only one User, we choose the OneToOne relationship.

Because we want to navigate from a user to a shop, we add the relation into the User class, we use the @OneToOne annotation.

We also adapt our DataLoader class, so that a user is already connected to a shop.

**tags**: #RelevantForExam